% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlR2-package.R
\docType{package}
\name{curlR2-package}
\alias{curlR2}
\alias{curlR2-package}
\title{curlR2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
