% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{EmailConfig_qq}
\alias{EmailConfig_qq}
\title{Authorize for QQ Email}
\usage{
EmailConfig_qq(user, PassCode)
}
\arguments{
\item{user}{the email address of the sender}

\item{PassCode}{the pass code of the sender, the length is 16 for QQ \code{PassCode}}
}
\description{
Configuration will be write to \code{curlR/config/config.json}, where user, PassCode
and server will be stored.
}
\section{QQ Email Authorization}{

\enumerate{
\item login in at \url{https://mail.qq.com/cgi-bin/frame_html};
\item setting -> get authorization \code{PassCode}
}
}

\examples{
\dontrun{
EmailConfig_qq("991810576@qq.com", "XXXXXXXX")
}
}
