% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_qq.R
\name{login_qq}
\alias{login_qq}
\alias{login_qq.pwd}
\alias{login_qq.app}
\title{login qq}
\usage{
login_qq.pwd(p, user = NULL, pwd = NULL)

login_qq.app(p)
}
\description{
\itemize{
\item \code{login_qq.pwd}: login by password
\item \code{login_qq.app}: quick login in by qq or Tim app
}
}
